Move C++ `Store` to its own file
I tried binding `wasmtime_store_epoch_deadline_callback` in C++ but I
don't know enough C++ to figure out the syntax for doing so.
